Game server: implemented the XP & coin events (and the corresponding +exp_add & +money_add commands, the announcements, and the notice updates -- though not sure how the latter is officially handled post-KE). Resolves #298.

+++ src/AIServer/AIUser.cpp
@@ -190,10 +190,10 @@ void CUser::SetPartyExp(int iNpcExp, int iLoyalty, int iPartyLevel, int iMan)
 	SendExp(iNpcExp, iLoyalty);
 }
 
-void CUser::SendExp(int iExp, int iLoyalty, int tType)
+void CUser::SendExp(int32 iExp, int32 iLoyalty, int tType)
 {
 	Packet result(AG_USER_EXP);
-	result << GetID() << uint16(iExp) << uint16(iLoyalty);
+	result << GetID() << iExp << iLoyalty;
 	g_pMain->Send(&result);   	
 }
